+++ packages/core/button/src/components/__tests__/visual-regression/buttonSnapshotTest.ts
@@ -3,7 +3,6 @@ import {
   takeScreenShot,
 } from '@atlaskit/visual-regression/helper';
 
-// add some comments
 describe('Snapshot Test', () => {
   it('ButtonAppearances should match snapshot', async () => {
     const url = getExampleUrl(