+++ packages/editor/editor-core/src/__tests__/integration/panel/paste-rich-text.ts
@@ -1,38 +1,44 @@
 import { BrowserTestCase } from '@atlaskit/webdriver-runner/runner';
-import Page from '@atlaskit/webdriver-runner/wd-wrapper';
 
 import {
   editable,
   getDocFromElement,
   fullpage,
   quickInsert,
-  clipboardHelper,
   clipboardInput,
   copyAsHTMLButton,
 } from '../_helpers';
+import {
+  goToEditorTestingExample,
+  mountEditor,
+} from '../../__helpers/testing-example-helpers';
+import { selectors } from './_utils';
 
 BrowserTestCase(
   'paste-rich-text.ts: Paste rich text into panel',
   { skip: ['ie', 'edge', 'safari'] },
   async client => {
-    const browser = new Page(client);
-    await browser.goto(clipboardHelper);
-    await browser.isVisible(clipboardInput);
-    await browser.type(
+    const page = await goToEditorTestingExample(client);
+
+    await page.isVisible(clipboardInput);
+    await page.type(
       clipboardInput,
       '<p>this is a link <a href="http://www.google.com">www.google.com</a></p><p>more elements with some <strong>format</strong></p><p>some addition<em> formatting</em></p>',
     );
-    await browser.click(copyAsHTMLButton);
+    await page.click(copyAsHTMLButton);
 
-    await browser.goto(fullpage.path);
-    await browser.waitForSelector(editable);
+    await mountEditor(page, {
+      appearance: fullpage.appearance,
+      allowPanel: true,
+    });
 
-    await browser.click(editable);
-    await quickInsert(browser, 'Panel');
+    await page.click(editable);
+    await quickInsert(page, 'Panel');
+    await page.waitForSelector(selectors.PANEL_EDITOR_CONTAINER);
 
-    await browser.paste(editable);
+    await page.paste(editable);
 
-    const doc = await browser.$eval(editable, getDocFromElement);
+    const doc = await page.$eval(editable, getDocFromElement);
     expect(doc).toMatchDocSnapshot();
   },
 );