+++ packages/elements/user-picker/src/__tests__/unit/_testUtils.tsx
@@ -5,7 +5,7 @@ import { User } from '../../types';
 export const renderProp = (
   wrapper: ShallowWrapper,
   renderProp: string,
-  ...args
+  ...args: any[]
 ) => {
   const prop = wrapper.prop(renderProp);
   if (prop && typeof prop === 'function') {