+++ packages/elements/user-picker/src/components/styles.ts
@@ -2,18 +2,17 @@ import { AVATAR_SIZES, BORDER_WIDTH } from '@atlaskit/avatar';
 import { colors } from '@atlaskit/theme';
 import memoizeOne from 'memoize-one';
 import { getAvatarSize } from './utils';
-import { Appearance } from '../types';
 
 export const BORDER_PADDING = 6;
 export const PLACEHOLDER_PADDING = 8;
 
-export const getStyles = memoizeOne((width, appearance: Appearance) => ({
-  menu: (css, state) => ({
+export const getStyles = memoizeOne((width: string | number) => ({
+  menu: (css: any, state: any) => ({
     ...css,
     width,
     minWidth: state.selectProps.menuMinWidth,
   }),
-  control: (css, state) => ({
+  control: (css: any, state: any) => ({
     ...css,
     width,
     borderColor: state.isFocused
@@ -44,7 +43,7 @@ export const getStyles = memoizeOne((width, appearance: Appearance) => ({
           : colors.N30,
     },
     padding: 0,
-    minHeight: appearance === 'compact' ? 32 : 44,
+    minHeight: state.selectProps.appearance === 'compact' ? 32 : 44,
     alignItems: 'stretch',
     maxWidth: '100%',
   }),
@@ -54,7 +53,7 @@ export const getStyles = memoizeOne((width, appearance: Appearance) => ({
     paddingLeft,
     paddingRight,
     ...css
-  }) => ({
+  }: any) => ({
     ...css,
     opacity: 0,
     transition: css.transition + ', opacity 150ms',
@@ -64,11 +63,14 @@ export const getStyles = memoizeOne((width, appearance: Appearance) => ({
       color: colors.R400,
     },
   }),
-  indicatorsContainer: css => ({
+  indicatorsContainer: (css: any) => ({
     ...css,
     paddingRight: 4,
   }),
-  valueContainer: ({ paddingTop, paddingBottom, position, ...css }, state) => ({
+  valueContainer: (
+    { paddingTop, paddingBottom, position, ...css }: any,
+    state: any,
+  ) => ({
     ...css,
     flexGrow: 1,
     padding: 0,
@@ -78,17 +80,17 @@ export const getStyles = memoizeOne((width, appearance: Appearance) => ({
     overflow: 'hidden',
     flexWrap: state.selectProps.isMulti ? 'wrap' : 'nowrap',
   }),
-  multiValue: css => ({
+  multiValue: (css: any) => ({
     ...css,
     borderRadius: 24,
   }),
-  multiValueRemove: css => ({
+  multiValueRemove: (css: any) => ({
     ...css,
     backgroundColor: 'transparent',
     '&:hover': { backgroundColor: 'transparent' },
   }),
-  placeholder: (css, state) => {
-    const avatarSize = getAvatarSize(appearance);
+  placeholder: (css: any, state: any) => {
+    const avatarSize = getAvatarSize(state.selectProps.appearance);
     return {
       ...css,
       paddingLeft: !state.selectProps.isMulti
@@ -97,6 +99,7 @@ export const getStyles = memoizeOne((width, appearance: Appearance) => ({
           2 * BORDER_WIDTH[avatarSize] +
           AVATAR_SIZES[avatarSize]
         : PLACEHOLDER_PADDING,
+      paddingTop: 2,
       paddingRight: 2,
       display: 'block',
       whiteSpace: 'nowrap',
@@ -106,15 +109,15 @@ export const getStyles = memoizeOne((width, appearance: Appearance) => ({
       margin: 0,
     };
   },
-  option: css => ({
+  option: (css: any) => ({
     ...css,
     overflow: 'hidden',
   }),
-  input: ({ margin, paddingTop, ...css }) => ({
+  input: ({ margin, ...css }: any) => ({
     ...css,
     display: 'flex',
     alignSelf: 'center',
-    paddingBottom: appearance === 'compact' ? 1 : 0,
+    paddingBottom: 1,
     paddingLeft: PLACEHOLDER_PADDING,
     '& input::placeholder': {
       /* Chrome, Firefox, Opera, Safari 10.1+ */