+++ packages/elements/user-picker/src/components/utils.ts
@@ -12,8 +12,11 @@ import {
   Value,
 } from '../types';
 
-export const isUser = (option: OptionData): option is User =>
-  option.type === undefined || option.type === UserType;
+export const isUser = (option: Value): option is User =>
+  isSingle(option) && (option.type === undefined || option.type === UserType);
+
+export const isSingle = (option: Value): option is OptionData =>
+  !Array.isArray(option);
 
 export const isTeam = (option: OptionData): option is Team =>
   option.type === TeamType;
@@ -35,9 +38,7 @@ export const extractOptionValue = (value: AtlaskitSelectValue) => {
 };
 
 export const isIterable = (
-  a:
-    | Promisable<OptionData | OptionData[]>
-    | Iterable<Promisable<OptionData | OptionData[]>>,
+  a: any,
 ): a is Iterable<Promisable<OptionData | OptionData[]>> =>
   typeof a[Symbol.iterator] === 'function';
 
@@ -45,7 +46,16 @@ export const getOptions = memoizeOne((options: OptionData[]) =>
   options.map(optionToSelectableOption),
 );
 
-export const optionToSelectableOptions = memoizeOne((defaultValue: Value) => {
+export interface OptionToSelectableOptions {
+  (defaultValue: OptionData): Option;
+  (defaultValue: OptionData[]): Option[];
+  (defaultValue?: null): null;
+  (defaultValue?: Value): Option | Option[] | null | undefined;
+}
+
+export const optionToSelectableOptions: OptionToSelectableOptions = memoizeOne<
+  OptionToSelectableOptions
+>(((defaultValue: Value) => {
   if (!defaultValue) {
     return null;
   }
@@ -53,7 +63,7 @@ export const optionToSelectableOptions = memoizeOne((defaultValue: Value) => {
     return defaultValue.map(optionToSelectableOption);
   }
   return optionToSelectableOption(defaultValue);
-});
+}) as OptionToSelectableOptions);
 
 export const getAvatarSize = (
   appearance: string,
@@ -73,7 +83,7 @@ export const hasValue = (value?: string): value is string =>
   !!value && value.trim().length > 0;
 
 export const callCallback = <U extends any[], R>(
-  callback: ((...U) => R) | undefined,
+  callback: ((...U: U) => R) | undefined,
   ...args: U
 ): R | undefined => callback && callback(...args);