+++ packages/media/media-avatar-picker/examples/2-image-navigator-tall.tsx
@@ -1,5 +1,6 @@
 /* tslint:disable:no-console */
 import * as React from 'react';
+import { IntlProvider } from 'react-intl';
 import { tallImage } from '@atlaskit/media-test-helpers';
 import ImageNavigator from '../src/image-navigator';
 
@@ -20,28 +21,30 @@ function handleImgRef(img: any) {
 }
 
 export default () => (
-  <div>
-    <h1>Local tall image</h1>
-    <ImageNavigator
-      imageSource={tallImage}
-      onPositionChanged={(x, y) => {
-        console.log('onPositionChanged', x, y);
-      }}
-      onSizeChanged={size => {
-        console.log('onSizeChanged', size);
-      }}
-      onLoad={onLoad}
-      onRemoveImage={() => console.log('onRemoveImage')}
-      onImageError={errorMessage => console.log('onImageError', errorMessage)}
-      onImageLoaded={(file, crop) => console.log('onImageLoaded', file, crop)}
-      onImageUploaded={file => console.log('onImageLoaded', file)}
-    />
-    <button onClick={exportImage}>Export</button>
-    <img
-      style={{ position: 'absolute', top: 0, left: '300px' }}
-      src=""
-      alt=""
-      ref={handleImgRef}
-    />
-  </div>
+  <IntlProvider locale="en">
+    <div>
+      <h1>Local tall image</h1>
+      <ImageNavigator
+        imageSource={tallImage}
+        onPositionChanged={(x, y) => {
+          console.log('onPositionChanged', x, y);
+        }}
+        onSizeChanged={size => {
+          console.log('onSizeChanged', size);
+        }}
+        onLoad={onLoad}
+        onRemoveImage={() => console.log('onRemoveImage')}
+        onImageError={errorMessage => console.log('onImageError', errorMessage)}
+        onImageLoaded={(file, crop) => console.log('onImageLoaded', file, crop)}
+        onImageUploaded={file => console.log('onImageLoaded', file)}
+      />
+      <button onClick={exportImage}>Export</button>
+      <img
+        style={{ position: 'absolute', top: 0, left: '300px' }}
+        src=""
+        alt=""
+        ref={handleImgRef}
+      />
+    </div>
+  </IntlProvider>
 );