+++ packages/media/media-avatar-picker/src/__tests__/unit/avatar-picker-dialogSpec.tsx
@@ -2,19 +2,19 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 import ModalDialog from '@atlaskit/modal-dialog';
 import Button from '@atlaskit/button';
-import { Avatar } from '../../../src/avatar-list';
-import { ImageNavigator, CONTAINER_SIZE } from '../../../src/image-navigator';
-import { PredefinedAvatarList } from '../../../src/predefined-avatar-list';
+import { smallImage, mountWithIntlContext } from '@atlaskit/media-test-helpers';
+import { Avatar } from '../../avatar-list';
+import { ImageNavigator, CONTAINER_SIZE } from '../../image-navigator';
+import { PredefinedAvatarList } from '../../predefined-avatar-list';
+import { AvatarPickerDialog } from '../../avatar-picker-dialog';
+import { DEFAULT_VISIBLE_PREDEFINED_AVATARS } from '../../avatar-picker-dialog/layout-const';
+import { dataURItoFile, fileToDataURI } from '../../util';
+import { PredefinedAvatarView } from '../../predefined-avatar-view';
 import {
-  AvatarPickerDialog,
+  Mode,
   AvatarPickerDialogProps,
   AvatarPickerDialogState,
-} from '../../../src/avatar-picker-dialog';
-import { DEFAULT_VISIBLE_PREDEFINED_AVATARS } from '../../../src/avatar-picker-dialog/layout-const';
-import { dataURItoFile, fileToDataURI } from '../../../src/util';
-import { smallImage, mountWithIntlContext } from '@atlaskit/media-test-helpers';
-import { PredefinedAvatarView } from '../../../src/predefined-avatar-view';
-import { Mode } from '../../../src';
+} from '../../avatar-picker-dialog/types';
 
 describe('Avatar Picker Dialog', () => {
   const renderWithProps = (props: Partial<AvatarPickerDialogProps>) => {