+++ packages/media/media-ui/src/InlineCard/UnauthorisedView/index.tsx
@@ -3,6 +3,7 @@ import { IconAndTitleLayout } from '../IconAndTitleLayout';
 import Button from '@atlaskit/button';
 import { truncateUrlForErrorView } from '../utils';
 import { Frame } from '../Frame';
+import { colors } from '@atlaskit/theme';
 
 export interface InlineCardUnauthorizedViewProps {
   /** The url to display */
@@ -31,7 +32,14 @@ export class InlineCardUnauthorizedView extends React.Component<
     const { url, icon, onClick, isSelected, onAuthorise } = this.props;
     return (
       <Frame onClick={onClick} isSelected={isSelected}>
-        <IconAndTitleLayout icon={icon} title={truncateUrlForErrorView(url)} />
+        <IconAndTitleLayout
+          icon={icon}
+          title={
+            <span style={{ color: colors.N500 }}>
+              {truncateUrlForErrorView(url)}
+            </span>
+          }
+        />
         {!onAuthorise ? (
           ''
         ) : (