+++ packages/editor/editor-jira-transformer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-jira-transformer
 
+## 5.2.4
+- [patch] [557a2b5734](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/557a2b5734):
+
+  - ED-5788: bump prosemirror-view and prosemirror-model
+
 ## 5.2.3
 - Updated dependencies [69c8d0c19c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/69c8d0c19c):
   - @atlaskit/editor-common@31.0.0